+++ libs/json/parse.c
@@ -8,10 +8,10 @@
 #include <json.h>
 #include <utils.h>
 
-static void parse_v(jsonelement_t *element, const char *text, const unsigned long long length, unsigned long long *i);
-static void parse_kv(jsonelement_t *parent, jsonelement_t **element, const char *text, const unsigned long long length, unsigned long long *i);
+static void parse_v(jsonelement_t *element, const char *text, const unsigned int length, unsigned int *i);
+static void parse_kv(jsonelement_t *parent, jsonelement_t **element, const char *text, const unsigned int length, unsigned int *i);
 
-static void parse_v(jsonelement_t *element, const char *text, const unsigned long long length, unsigned long long *i) {
+static void parse_v(jsonelement_t *element, const char *text, const unsigned int length, unsigned int *i) {
 	bool escaping = false;
 
 	while (*i < length) {
@@ -158,7 +158,7 @@ static void parse_v(jsonelement_t *element, const char *text, const unsigned lon
 					((double *) element->value)[0] = ((((double *) element->value)[0] * 10) + (ch - 48));
 				} else if (ch == '.') {
 					++(*i);
-					unsigned long starts_at = *i;
+					unsigned int starts_at = *i;
 
 					bool condition = (*i < length);
 
@@ -185,8 +185,8 @@ static void parse_v(jsonelement_t *element, const char *text, const unsigned lon
 	}
 }
 
-static void parse_kv(jsonelement_t *parent, jsonelement_t **element, const char *text, const unsigned long long length, unsigned long long *i) {
-	unsigned long key_length = 0;
+static void parse_kv(jsonelement_t *parent, jsonelement_t **element, const char *text, const unsigned int length, unsigned int *i) {
+	unsigned int key_length = 0;
 	bool parsing_key = false;
 	bool parsing_value = false;
 
@@ -243,8 +243,8 @@ static void parse_kv(jsonelement_t *parent, jsonelement_t **element, const char 
 
 jsonelement_t *json_parse(const char *text) {
 	jsonelement_t *result = allocate(NULL, -1, 1, sizeof(jsonelement_t));
-	const unsigned long long length = strlen(text);
-	unsigned long long i = 0;
+	const unsigned int length = strlen(text);
+	unsigned int i = 0;
 
 	parse_v(result, text, length, &i);
